dft: implemented dft

+++ inc/dft.h
@@ -3,7 +3,7 @@
  * @author reusa1 (reusa1@bfh.ch)
  * @brief Implementation of calculating a discrete fourier transform.
  * @version 0.1
- * @date 2022-01-07
+ * @date 2022-01-14
  * 
  * @copyright Copyright (c) 2022 Adrian Reusser
  * 
@@ -28,9 +28,13 @@
 
 /**
  * @brief 
+ * calculates magnitudes for the spectogram
+ * uses first 60 data points of 1024
+ * uses one channel to calculate
+ * cos is stored as an array of parameters so calculation can bea avoided 
  * 
  * @param samples 
  * @param magnitude 
  * @retval
  */
-int dft_transform(int16_t samples[], uint16_t magnitude[]);
+int dft_transform(int16_t samples[], uint32_t magnitude[]);
